Reduce unnecessary measure
If a window size can be known by the LayoutParams, we can use the size
to relayout window, and we don't have to measure the view hierarchy
before laying out the views.

The CL applies the optimization on notification shade since it takes
much time on performing measure. We plan to apply this on all the
windows later.

Fix: 259095271
Fix: 259673244
Bug: 260382739
Test: Change display density via adb command and see if notification
      shade gets multiple measure in one traversal.
Change-Id: Ib85ceeda8a821703a32e88de7c67999f63276b6b