Fix rendering issue after cancelling Notifications inside a group
If we return true from `offerToKeepInParentForAnimation`, the
ShadeViewDiffer won't detach this row and the view system is responsible
for ensuring the row is in eventually removed from the parent. The view
system can only guarantee this if the removal was triggered by a dismissal from the user.

The bug was a regression from  ag/20291263.

Fixes: 260079710
Test: 1) post a group of notifications 2) cancel a child 3) post the
same child again (a notif with the same tag) 4) expand notif group
Test: dismiss a group of Notifications and observe the animations
Test: atest ExpandableNotificationRowControllerTest

Change-Id: Iea4032ece37e36a3f08cd6626e998717cda511c7