Let DesktopModeController reorder tasks on TO_FRONT transitions
When a fullscreen task is moved to the front of the freeform TDA from
recents, any previously visible freeform apps should remain active
in front of home and behind the task that is moving to front. Since
they were sent behind everything when recents opened, they need to be
moved to the front again by the controller. Let #handleRequest accept
TRANSIT_TO_FRONT transitions as a signal to move desktop apps to the
front, just like it does for TRANSIT_OPEN.

Bug: 259453661
Test: atest DesktopModeControllerTest
Test: manual -
 1. Turn desktop mode off, close all apps from recents
 2. Launch Gmail
 3. Launch Photos
 4. Turn desktop mode on from quick settings
 5. Observe Photos is freeform and visible, Gmail is invisible behind
 home
 6. Swipe up to recents, tap on Gmail tile
 7. Observe Gmail and Photos are both visible in freeform over home
Change-Id: I826b926dbc572b48a38d4c49dd4153586ee11469