Prepare ActivityInterceptorCallback to be unhidden to mainline modules.
AdServices mainline module need to register ActivityInterceptorCallback
as part of Activity Support in SDK Sandbox.

ActivityInterceptorCallback should be visible to mainline modules which
requires the package com.android.server.wm to tagged with Hide annotation,
that is why in this CL a file packge-info.java is added to the package.

This also add fixed to make the class follow go/android-api-guidelines

Test: atest com.android.server.wm.ActivityStartInterceptorTest && atest com.android.server.wm.ActivityTaskManagerServiceTests
Bug: 248531721
CTS-Coverage-Bug: 261596770
API-Coverage-Bug: 261596770
Change-Id: Ied9787c1fa691d609ee7625433dc088d79d11473