Remove a hack to fix a short power button press behavior.
A hack was introduced in ag/20583933 to fix an issue with short power
button press when not the primary user (basically, in this case, the
screen should turn off but instead nothing happened). The underlying
issue has been fixed, so this change removes the hack.

Bug: 260135090
Test: manually by setting up a secondary user and then verifying that
the device will screen off on power button press.

Change-Id: Ia1303dea4c3b72f7b72e069912286e6e9666b491