Ensure proguard doesn't over-optimize Flags Library.
In the main build, proguard is optimizing out Parcelable flags, which
means that clients of the library crash when they receive those flags
over the wire.

Telling proguard to keep Parcelable flags makes everything happy.

This fix is really for main, but the behavior is desired in qpr as
well, (even though proguard doesn't cause a problem in qpr).

Fixes: 254310572
Test: manually built and run on main/master
Change-Id: I541c405cd5642fae7c0cbdbf3ddcffd49e19eed1