Fix a11y focus rendering on non-dirty root view.
ViewRootImpl can receive TYPE_VIEW_ACCESSIBILITY_FOCUSED, which
means it should draw the a11y focus. But without any other
triggers (e.g. animation or scrolling) it is not being rendered.
This event invalidates the renderer, which makes the focus rectangle
dirty, but nothing make the draw() method of the view to be called.
Fixed by adding the whole view invalidation.

Fix: 264356970
Test: manual
Change-Id: I4073974551daf622a0c467bfb27df5f5a14ba82c