Use ALSA for USB MIDI 1.0 devices
In Android T, we added USB MIDI 2.0 support. As part of the change,
we changed USB MIDI 1.0 devices to also not use ALSA.

Users have reported bugs where audio stops working after MIDI is used.
The root cause of this is because ALSA disconnects the sound card if
one part of the sound card is removed. Android disconnects the ALSA
MIDI interface as a direct USB connection is made.

As long as ALSA is used for audio, ALSA needs to be used for MIDI.
Thus, Android will go back to using ALSA for MIDI 1.0 USB devices.

Refer to ag/16835029 for the change where ALSA is no longer used for
host mode. This change is a partial revert of it.

Bug: 251243033
Bug: 259365701
Test: MIDI Scope on Korg device
Test: MIDI Keyboard and Scope loopback on 2x2 MIDI interface
Test: MIDI Scope on MIDI 2.0 device
Test: OboeTester Tap to Tone with MIDI and Audio
Change-Id: Ia6a6021619e05dc5b483d4644d7fd152d89e1d46
(cherry picked from commit 8e75a203179078e04f1e3fe7ddbee849908f4b3f)