FBE notification is updated when language changed
The FBE(encryption) notification shown when the work profile
is locked still consists of the previous language even if
the language is changed. The following commit resolve that issue.
When the language is changed, the notification is re-configured
in the changed language. After receiving 'ACTION_LOCALE_CHANGED'
broadcast, the notification will be updated in the changed language.

Manual testing
* Create work profile
* Set work profile lock type
* Evict work profile's ce key or reboot.
* Language change

Bug: 239797643
Test: manual testing
Change-Id: I1c97295d1053a2cabea33959bc1dbf5ed88173a1
Signed-off-by: seokil <seokil.kang@samsung.corp-partner.google.com>